add processor utils with lazy imports and no transformers to use in environment
