+++ verifiers/utils/image_utils.py
@@ -9,3 +9,12 @@ def _base64_to_pil(data_uri: str) -> Image.Image:
     header, b64data = data_uri.split(",", 1)
     image_data = base64.b64decode(b64data)
     return Image.open(BytesIO(image_data)).convert("RGB")
+
+def pil_to_base64_url(pil_image) -> str:
+    """
+    Convert a PIL image to a base64 URL string suitable for OpenAI/vLLM messages.
+    """
+    buffered = BytesIO()
+    pil_image.save(buffered, format="PNG")
+    img_str = base64.b64encode(buffered.getvalue()).decode("utf-8")
+    return f"data:image/png;base64,{img_str}"
